+++ tests/test_daterange.py
@@ -33,7 +33,7 @@ def test_from_date():
     (date(2000, 1, 3), daterange(date(2000, 1, 3), date(2000, 1, 10))),
 ])
 def test_from_date_week(day, span):
-    assert daterange.from_date(day, what="week") == span
+    assert daterange.from_date(day, period="week") == span
 
 
 @pytest.mark.parametrize("day, span", [
@@ -42,7 +42,7 @@ def test_from_date_week(day, span):
     (date(2000, 1, 3), daterange(date(2000, 1, 2), date(2000, 1, 9))),
 ])
 def test_from_date_american_week(day, span):
-    assert daterange.from_date(day, what="american_week") == span
+    assert daterange.from_date(day, period="american_week") == span
 
 
 @pytest.mark.parametrize("day, span", [
@@ -51,7 +51,7 @@ def test_from_date_american_week(day, span):
     (date(2001, 2, 15), daterange(date(2001, 2, 1), date(2001, 2, 28), upper_inc=True)),
 ])
 def test_from_date_month(day, span):
-    assert daterange.from_date(day, what="month") == span
+    assert daterange.from_date(day, period="month") == span
 
 
 @pytest.mark.parametrize("day, span", [
@@ -60,7 +60,7 @@ def test_from_date_month(day, span):
     (date(2000, 3, 31), daterange(date(2000, 1, 1), date(2000, 3, 31), upper_inc=True)),
 ])
 def test_from_date_quarter(day, span):
-    assert daterange.from_date(day, what="quarter") == span
+    assert daterange.from_date(day, period="quarter") == span
 
 
 @pytest.mark.parametrize("day, span", [
@@ -69,7 +69,7 @@ def test_from_date_quarter(day, span):
     (date(2000, 12, 31), daterange(date(2000, 1, 1), date(2001, 1, 1))),
 ])
 def test_from_date_year(day, span):
-    assert daterange.from_date(day, what="year") == span
+    assert daterange.from_date(day, period="year") == span
 
 
 @pytest.mark.parametrize("param", [
@@ -81,6 +81,47 @@ def test_from_date_year(day, span):
 def test_from_date_type_check(param):
     with pytest.raises(TypeError):
         daterange.from_date(param)
+
+
+@pytest.mark.parametrize("year, iso_week, first_day", [
+    (2000, 1, date(2000, 1, 3)),
+    (2000, 2, date(2000, 1, 10)),
+    (2000, 3, date(2000, 1, 17)),
+    (2000, 4, date(2000, 1, 24)),
+])
+def test_from_week(year, iso_week, first_day):
+    assert daterange.from_week(year, iso_week) == daterange.from_date(
+        first_day, period="week")
+
+
+@pytest.mark.parametrize("year", [
+    2000,
+    2001,
+])
+@pytest.mark.parametrize("month", range(1, 13))
+def test_from_month(year, month):
+    assert daterange.from_month(year, month) == daterange.from_date(
+        date(year, month, 1), period="month")
+
+
+@pytest.mark.parametrize("year, quarter, first_day", [
+    (2000, 1, date(2000, 1, 1)),
+    (2000, 2, date(2000, 4, 1)),
+    (2000, 3, date(2000, 7, 1)),
+    (2000, 4, date(2000, 10, 1)),
+])
+def test_from_quarter(year, quarter, first_day):
+    assert daterange.from_quarter(year, quarter) == daterange.from_date(
+        first_day, period="quarter")
+
+
+@pytest.mark.parametrize("year", [
+    2000,
+    2001,
+])
+def test_from_year(year):
+    assert daterange.from_year(year) == daterange.from_date(
+        date(year, 1, 1), period="year")
 
 
 def test_last():